Cleaned autoloads.
Remove existing autoloads. Add an autoload for 'django-el-mode'.

Not sure how lazy loading will affect the keymap, from what I
understand from the docs it seems ok to mark it as autoload, but I've
preferred to change the way the keymap is bound to a prefix key. The
user configuration may assign a value to 'django-el-prefix-key' before
activating the mode.