Save position before jumping to a template/js controller.
The commands for jumping to the template or js controller have
semantics resembling those of the jump to definition command so it
seems natural to be able to get back to the original place by using
"xref-pop-marker-stack".

In order to acomplish that they push the mark into xref's marker
stack.